+++ packages/rocketh-deploy-proxy/src/index.ts
@@ -5,6 +5,9 @@ import type {EIP1193Account} from 'eip-1193';
 import {extendEnvironment} from 'rocketh';
 import {Chain, DeployContractParameters, encodeAbiParameters, encodeDeployData, encodeFunctionData} from 'viem';
 import artifactsAsModule from 'solidity-proxy/generated/artifacts';
+import {logs} from 'named-logs';
+
+const logger = logs('rocketh-deploy-proxy');
 
 // fix for weird loading issue
 const artifacts = (artifactsAsModule as any).default
@@ -14,16 +17,25 @@ const artifacts = (artifactsAsModule as any).default
 export type ProxyDeployOptions = {
 	disabled?: boolean;
 	owner: EIP1193Account;
-	execute?: string; // TODO
-	deployImplementation?: <TAbi extends Abi, TChain extends Chain>(
-		name: string,
-		args: DeploymentConstruction<TAbi, TChain>
-	) => Promise<Deployment<TAbi>>;
+	execute?: string;
+};
+
+export type ImplementationDeployer<TAbi extends Abi, TChain extends Chain> = (
+	name: string,
+	args: Omit<DeploymentConstruction<TAbi, TChain>, 'artifact'>
+) => Promise<Deployment<TAbi>>;
+
+// TODO omit nonce ? // TODO omit chain ? same for rocketh-deploy
+export type ProxyEnhancedDeploymentConstruction<TAbi extends Abi, TChain extends Chain = Chain> = Omit<
+	DeploymentConstruction<TAbi, TChain>,
+	'artifact'
+> & {
+	artifact: string | Artifact<TAbi> | ImplementationDeployer<TAbi, TChain>;
 };
 
 export type DeployViaProxyFunction = <TAbi extends Abi, TChain extends Chain = Chain>(
 	name: string,
-	args: DeploymentConstruction<TAbi, TChain>,
+	args: ProxyEnhancedDeploymentConstruction<TAbi, TChain>,
 	options?: ProxyDeployOptions
 ) => Promise<Deployment<TAbi>>;
 
@@ -36,15 +48,13 @@ declare module 'rocketh' {
 extendEnvironment((env: Environment) => {
 	async function deployViaProxy<TAbi extends Abi, TChain extends Chain = Chain>(
 		name: string,
-		args: DeploymentConstruction<TAbi>, // TODO omit nonce ? // TODO omit chain ? same for rocketh-deploy
+		args: ProxyEnhancedDeploymentConstruction<TAbi, TChain>,
 		options?: ProxyDeployOptions
 	): Promise<Deployment<TAbi>> {
 		const proxyName = `${name}_Proxy`;
 		const implementationName = `${name}_Implementation`;
 
 		const {account, artifact, ...viemArgs} = args;
-		// TODO throw specific error if artifact not found
-		const artifactToUse = (typeof artifact === 'string' ? env.artifacts[artifact] : artifact) as Artifact<TAbi>;
 
 		let address: `0x${string}`;
 		if (account.startsWith('0x')) {
@@ -62,34 +72,50 @@ extendEnvironment((env: Environment) => {
 
 		const proxyArtifact = artifacts.EIP173Proxy;
 
-		const implementation = options?.deployImplementation
-			? await options?.deployImplementation(implementationName, {...args})
-			: await env.deploy(implementationName, {
-					...args,
-			  });
+		const implementation =
+			typeof args.artifact === 'function'
+				? await args.artifact(implementationName, {...args})
+				: await env.deploy(implementationName, {
+						...args,
+				  } as DeploymentConstruction<TAbi, TChain>);
+
+		logger.info(`implementation at ${implementation.address}`, `${implementationName}`);
+
+		const {
+			address: implementationAddress,
+			argsData: implementationArgsData,
+			transaction,
+			...artifactFromImplementation
+		} = implementation;
+
+		// TODO throw specific error if artifact not found
+		const artifactToUse =
+			typeof args.artifact === 'function'
+				? artifactFromImplementation
+				: ((typeof artifact === 'string' ? env.artifacts[artifact] : artifact) as Artifact<TAbi>);
 
 		let existingDeployment = env.get<TAbi>(name);
 
-		const owner = options?.owner || address;
+		logger.info(`existingDeployment at ${existingDeployment?.address}`);
 
-		// TODO better way to get args Data :D
-		const bytecode = artifactToUse.bytecode;
-		const abi = artifactToUse.abi;
-		const argsToUse: DeployContractParameters<TAbi, TChain> = {
-			...viemArgs,
-			account,
-			abi,
-			bytecode,
-		} as unknown as DeployContractParameters<TAbi, TChain>; // TODO why casting necessary here
-		const calldata = encodeDeployData(argsToUse);
-		const argsData = `0x${calldata.replace(bytecode, '')}` as `0x${string}`;
+		const owner = options?.owner || address;
 
 		let methodCallData: `0x${string}` | undefined;
 		if (options?.execute) {
-			const method: AbiFunction | undefined = abi.find(
+			const method: AbiFunction | undefined = artifactToUse.abi.find(
 				(v) => v.type === 'function' && v.name === options.execute
 			) as AbiFunction;
 			if (method) {
+				// TODO better way to get args Data :D
+				const bytecode = artifactToUse.bytecode;
+				const abi = artifactToUse.abi;
+				const argsToUse: DeployContractParameters<TAbi, TChain> = {
+					...viemArgs,
+					account,
+					abi,
+					bytecode,
+				} as unknown as DeployContractParameters<TAbi, TChain>; // TODO why casting necessary here
+
 				methodCallData = encodeFunctionData({abi: [method], functionName: method.name, args: argsToUse as any});
 			}
 		}
@@ -101,10 +127,14 @@ extendEnvironment((env: Environment) => {
 				args: [implementation.address, owner, methodCallData ? methodCallData : '0x'], // TODO upgradeToAndCall argsData],
 			});
 
+			logger.info(`proxy deployed at ${proxy.address}`);
+
 			existingDeployment = await env.save<TAbi>(name, {
 				...proxy,
 				...artifactToUse,
 			});
+
+			logger.info(`saving as ${name}`);
 		} else {
 			const proxyDeployment = env.get(proxyName);
 			if (!proxyDeployment) {
@@ -118,6 +148,10 @@ extendEnvironment((env: Environment) => {
 			const currentImplementationAddress = `0x${implementationSlotData.substr(-40)}`;
 
 			if (currentImplementationAddress.toLowerCase() !== implementation.address.toLowerCase()) {
+				logger.info(
+					`different implementation old: ${currentImplementationAddress} new: ${implementation.address}, upgrade...`
+				);
+
 				if (methodCallData) {
 					await env.execute<typeof proxyArtifact.abi, 'upgradeToAndCall'>(proxyName, {
 						account: address,
@@ -137,6 +171,7 @@ extendEnvironment((env: Environment) => {
 				...proxyDeployment,
 				...artifactToUse,
 			});
+			logger.info(`saving as ${name}`);
 		}
 		return existingDeployment;
 	}